update users and user history and info

+++ src/main/java/liv/codveda/blog/app/repository/UsersRepository.java
@@ -8,7 +8,7 @@ import java.util.Optional;
 
 
 @Repository
-public interface UsersRepository extends JpaRepository<Users, Integer> {
+public interface UsersRepository extends JpaRepository<Users, Long> {
 
     Optional<Users> findByEmail(String email);
     boolean existsByEmail(String email);
